Update batch writer, add to opensearch sink.
Not using the batch API for opensearch (and likely for elasticsearch as well)
is very bad for perfomance, easliy causing event backups in moderately busy
clusters. This commit updates the batch writer to make it a little more clear
in naming and configuration (while also updating the one existing user,
bigquery, to keep its current interface). This also adds the ability to force
the opensearch sink to output events with a consistent field as their most
recent timestamp, which is very useful for datastreams.